Porting some changes to tests for Fedora 4 back to this branch
